requirements.txt
streamlit
scikit-learn
tensorflow-cpu
joblib
numpy
pandas